// seo-app/src/types.ts
export interface MediaItem {
  id: string;
  type: 'image' | 'video';
  url: string;
  name: string;
  altText: string;
  size?: number;
  mimeType?: string;
  history?: {
    altText?: string;
    updatedAt: string;
  }[];
  createdAt?: any;
}